Final Commit
Ready for review. Renamed the files to their proper names. Added in the variable to store the date. Added in a variable to store the date in string form. Added a variable to store the current day. Added a variable to store the html value. Inserts in the date and html after the content in an element.